+++ ChemFlow/src/ScoreFlow_functions.bash
@@ -175,16 +175,14 @@ ScoreFlow_rescore_vina() {
 #
 #        Author: Dona de Francquen
 #
-#    PARAMETERS: ${mgltools_folder}
-#                ${RUNDIR}
+#    PARAMETERS: ${RUNDIR}
 #                ${LIGAND_LIST}
 #                ${DOCK_CENTER}
 #                ${DOCK_LENGHT}
 #===============================================================================
 # Prepare RECEPTOR
 if [ ! -f ${RUNDIR}/receptor.pdbqt ] ; then
-    ${mgltools_folder}/bin/python \
-    ${mgltools_folder}/MGLToolsPckgs/AutoDockTools/Utilities24/prepare_receptor4.py \
+        pythonsh $(command -v prepare_receptor4.py) \
         -r ${RUNDIR}/receptor.mol2 \
         -o ${RUNDIR}/receptor.pdbqt
 fi
@@ -193,9 +191,8 @@ fi
 for LIGAND in ${LIGAND_LIST[@]} ; do
     # Prepare Ligands
     if [ ! -f ${RUNDIR}/${LIGAND}/ligand.pdbqt ] ; then
-        ${mgltools_folder}/bin/python \
-        ${mgltools_folder}/MGLToolsPckgs/AutoDockTools/Utilities24/prepare_ligand4.py \
-            -l ${RUNDIR}/${LIGAND}/ligand.mol2 \
+	    pythonsh $(command -v prepare_ligand4.py) \
+	    -l ${RUNDIR}/${LIGAND}/ligand.mol2 \
             -o ${RUNDIR}/${LIGAND}/ligand.pdbqt
     fi
     # Run vina
@@ -251,6 +248,10 @@ if [ ${WRITE_ONLY} == 'yes' ] ; then
 else
     for LIGAND in ${LIGAND_LIST[@]} ; do
     cd ${RUNDIR}/${LIGAND}
+
+        if [ ! -f ligand.frcmod ] ; then
+            parmchk2 -i ligand_gas.mol2 -o ligand.frcmod -f mol2 -s 2
+        fi
 
         if [ ${RUN_ONLY} == 'yes' ] ; then
             echo -e "cd ${RUNDIR}/${LIGAND}" > ScoreFlow.run
@@ -330,7 +331,7 @@ if [ ${CHARGE} != 'gas' ] ; then
         if [ "$(grep ${LIGAND_NAME} ${CHEMFLOW_HOME}/ChemBase/${CHARGE}/ChemBase_${CHARGE}.lst)" == ${LIGAND_NAME} ] ; then
             echo "${CHARGE} charges found in ChemBase for ${LIGAND}"
 
-            awk -v LIGAND=${LIGAND_NAME} '$0 ~ LIGAND {flag=1;next}/BOND/{flag=0}flag' ${CHEMFLOW_HOME}/ChemBase/${CHARGE}/ChemBase_${CHARGE}.mol2 | awk '/1 MOL/&&!/TEMP/ {print $9}' > charges.dat
+            awk -v LIGAND=${LIGAND_NAME} '$0 ~ LIGAND {flag=1;next}/BOND/{flag=0}flag' ${CHEMFLOW_HOME}/ChemBase/${CHARGE}/ChemBase_${CHARGE}.mol2 | awk '/ MOL/&&!/TEMP/ {print $9}' > charges.dat
             antechamber -i ligand_gas.mol2 -o ligand_${CHARGE}.mol2 -fi mol2 -fo mol2 -cf charges.dat -c rc -pf yes &> /dev/null
 
             # Done
@@ -342,8 +343,8 @@ if [ ${CHARGE} != 'gas' ] ; then
     if [ "${DONE_CHARGE}" == "false" ] && [ -f ${WORKDIR}/${PROJECT}.chemflow/LigFlow/${CHARGE}/${LIGAND_NAME}.mol2 ] ; then
         echo "${CHARGE} charges found in LigFlow for ${LIGAND}"
 
-        awk '/1 MOL/&&!/TEMP/ {print $9}' ${WORKDIR}/${PROJECT}.chemflow/LigFlow/${CHARGE}/${LIGAND_NAME}.mol2 > charges.dat
-        antechamber -i ligand_gas.mol2 -o ligand_${CHARGE}.mol2 -fi mol2 -fo mol2 -cf charges.dat -c rc -pf yes &> /dev/null
+        awk '/ MOL/&&!/TEMP/ {print $9}' ${WORKDIR}/${PROJECT}.chemflow/LigFlow/${CHARGE}/${LIGAND_NAME}.mol2 > charges.dat
+        antechamber -i ligand_gas.mol2 -o ligand_${CHARGE}.mol2 -fi mol2 -fo mol2 -cf charges.dat -c rc -pf yes -dr no &> /dev/null
 
         # Done
         DONE_CHARGE="true"
@@ -441,8 +442,8 @@ for filename in ${scoreflow_protocol} ; do
     eval echo \""${file}"\" > ${RUNDIR}/${filename}.in
 done
 
-#mmgbsa input
-echo "$(cat ${CHEMFLOW_HOME}/templates/mmgbsa/GB2.template)" > ${RUNDIR}/GB2.in
+#mm(pb,gb)sa input
+echo "$(cat ${CHEMFLOW_HOME}/templates/mmgbsa/${SCORING_FUNCTION}.template)" > ${RUNDIR}/${SCORING_FUNCTION}.in
 }
 
 
@@ -516,9 +517,14 @@ done" >> ${RUNDIR}/ScoreFlow.run
 
 if [ ! -f MMPBSA.dat ] ; then
 echo "rm -rf com.top rec.top ligand.top
-ante-MMPBSA.py -p ${init}.prmtop -c com.top -r rec.top -l ligand.top -n :MOL -s ':WAT,Na+,Cl-' --radii=mbondi2 &> ante_mmpbsa.job
-MMPBSA.py -O -i ../GB2.in -sp ${init}.prmtop -cp com.top -rp rec.top -lp ligand.top -o MMPBSA.dat -eo MMPBSA.csv -y ${TRAJECTORY} &> MMPBSA.job
-rm -rf reference.frc " >> ${RUNDIR}/ScoreFlow.run
+ante-MMPBSA.py -p ${init}.prmtop -c com.top -r rec.top -l ligand.top -n :MOL -s ':WAT,Na+,Cl-' --radii=mbondi2 &> ante_mmpbsa.job" >>${RUNDIR}/ScoreFlow.run
+
+if [ "${WATER}" != "yes" ] ; then
+echo "MMPBSA.py -O -i ../mmgbsa.in -cp com.top -rp rec.top -lp ligand.top -o MMPBSA.dat -eo MMPBSA.csv -y ${TRAJECTORY} &> MMPBSA.job" >>${RUNDIR}/ScoreFlow.run
+else
+echo "MMPBSA.py -O -i ../mmgbsa.in -sp ${init}.prmtop -cp com.top -rp rec.top -lp ligand.top -o MMPBSA.dat -eo MMPBSA.csv -y ${TRAJECTORY} &> MMPBSA.job" >>${RUNDIR}/ScoreFlow.run
+fi
+echo "rm -rf reference.frc " >> ${RUNDIR}/ScoreFlow.run
 fi
 }
 
@@ -639,7 +645,7 @@ if [ -f ${RUNDIR}/ScoreFlow.csv ]  ; then
     rm -rf ${RUNDIR}/ScoreFlow.csv
 fi
 
-SCOREFLOW_HEADER="DOCK_PROGRAM PROTOCOL RECEPTOR LIGAND POSE SCORE"
+SCOREFLOW_HEADER="SCORE_PROGRAM PROTOCOL RECEPTOR LIGAND POSE SCORE"
 
 case ${SCORE_PROGRAM} in
 "PLANTS")
@@ -750,13 +756,16 @@ JOB SCHEDULLER: ${JOB_SCHEDULLER}
      OVERWRITE: ${OVERWRITE}
 "
 
-echo -n "
-Continue [y/n]? "
-read opt
-case $opt in
-"Y"|"YES"|"Yes"|"yes"|"y")  ;;
-*)  echo "Exiting" ; exit 0 ;;
-esac
+if [ "${YESTOALL}" != 'yes' ] ; then
+  echo -n "
+  Continue [y/n]? "
+  read opt
+  case $opt in
+  "Y"|"YES"|"Yes"|"yes"|"y")  ;;
+  *)  echo "Exiting" ; exit 0 ;;
+  esac
+fi
+
 }
 
 
@@ -815,18 +824,18 @@ ScoreFlow -r receptor.mol2 -l ligand.mol2 -p myproject --center X Y Z [--protoco
 # For MMGBSA only
 ScoreFlow -r receptor.pdb -l ligand.mol2 -p myproject [-protocol protocol-name] -sf mmgbsa
 
-[Help]
+[ Help ]
  -h/--help              : Show this help message and quit
  -hh/--fullhelp         : Detailed help
 
-[Required]
-*-p/--project       STR : ChemFlow project
-*-r/--receptor     FILE : Receptor MOL2 file
-*-l/--ligand       FILE : Ligands  MOL2 file
+[ Required ]
+ -p/--project       STR : ChemFlow project
+ -r/--receptor     FILE : Receptor MOL2 file
+ -l/--ligand       FILE : Ligands  MOL2 file
 
-[Optional]
+[ Optional ]
  --protocol         STR : Name for this specific protocol [default]
- -sf/--function     STR : vina, chemplp, plp, plp95, mmgbsa [chemplp]
+ -sf/--function     STR : vina, chemplp, plp, plp95, mmgbsa, mmpbsa [chemplp]
 
 [ Charges for ligands - MMGBSA ]
  --gas                  : Gasteiger-Marsili (default)
@@ -849,16 +858,22 @@ ScoreFlow -r receptor.pdb -l ligand.mol2 -p myproject [-protocol protocol-name] 
  --overwrite            : Overwrite results
 
 [ Rescoring with vina or plants ]
-Note: You can automatically get the center and radius/size for a particular ligand .mol2 file by using the ${CHEMFLOW_HOME}/bin/bounding_shape.py script
-*--center           STR : xyz coordinates of the center of the binding site, separated by a space
-[ PLANTS ]
+
+ --center           STR : xyz coordinates of the center of the binding site, separated by a space
+
+ [ PLANTS ]
  --radius         FLOAT : Radius of the spheric binding site [15]
-[ Vina ]
+
+ [ Vina ]
  --size            LIST : Size of the grid along the x, y and z axis, separated by a space [15 15 15]
  --vina-mode        STR : local_only (local search then score) or score_only [local_only]
 
 [ Post Processing ]
  --postprocess          : Process ScoreFlow output for the specified project/protocol/receptor.
+
+Note: You can automatically get the center and radius/size 
+      for a particular ligand .mol2 file using the bounding_shape.py script
+
 _________________________________________________________________________________
 "
 exit 0
@@ -986,6 +1001,12 @@ while [[ $# -gt 0 ]]; do
         #    --archive)
         #      ARCHIVE='yes'
         #    ;;
+        "--yes")
+            YESTOALL='yes'
+        ;;
+        "--cuda-double")
+            CUDA_PRECISION="DOUBLE"
+        ;;      
         *)
             unknown="$1"        # unknown option
             echo "Unknown flag \"$unknown\". RTFM"