+++ docs/history.rst
@@ -1,6 +1,19 @@
+.. highlight:: bash
+
 =======
 History
 =======
+1.0 (2019-10-23)
+----------------
+* ChemFlow community release !
+
+0.92 (2019-08-21)
+----------------
+* ChemFlow community pre-release !
+
+0.91 (2019-02-05)
+-----------------
+* Documentation update
 
 0.9 (2018-07-27)
 ------------------